+++ BookManager/AddBookViewController.swift
@@ -1,25 +1,25 @@
 import UIKit
 import APIKit
 
-class AddBookViewController: BaseBookViewController, FileAttachable {
+class AddBookViewController: UIViewController, FileAttachable, BookValidateable {
 
     @IBOutlet weak var imageView: UIImageView!
 
-    @IBOutlet override weak var nameTextField: UITextField! {
+    @IBOutlet weak var nameTextField: UITextField! {
         didSet {
             nameTextField.apply(borderWidth: Const.TextFieldBorderWidth, borderColor: Const.TextFieldBorderColor,
                               radius: Const.TextFieldCornerRadius, masksToBound: Const.TextFieldMasksToBounds)
         }
     }
 
-    @IBOutlet override weak var priceTextField: UITextField! {
+    @IBOutlet weak var priceTextField: UITextField! {
         didSet {
             priceTextField.apply(borderWidth: Const.TextFieldBorderWidth, borderColor: Const.TextFieldBorderColor,
                                radius: Const.TextFieldCornerRadius, masksToBound: Const.TextFieldMasksToBounds)
         }
     }
 
-    @IBOutlet override weak var purchaseDateTextField: UITextField! {
+    @IBOutlet weak var purchaseDateTextField: UITextField! {
         didSet {
             purchaseDateTextField.apply(borderWidth: Const.TextFieldBorderWidth,
                                         borderColor: Const.TextFieldBorderColor,
@@ -47,7 +47,8 @@ class AddBookViewController: BaseBookViewController, FileAttachable {
     }
 
     @IBAction func didSaveButtonTapped(_ sender: Any) {
-        let errors = validate()
+        let errors = validate(name: nameTextField.text!, price: priceTextField.text!)
+
         if errors.count > 0 {
             return AlertHelper.showAlert(self, title: R.string.localizable.validateErrorTitle(),
                                          message: errors.joined(separator: "\n"))
@@ -83,6 +84,10 @@ class AddBookViewController: BaseBookViewController, FileAttachable {
 
     override func didReceiveMemoryWarning() {
         super.didReceiveMemoryWarning()
+    }
+
+    func pickerChanged(sender: UIDatePicker) {
+        DatePickerHelper.setValue(sender, target: purchaseDateTextField)
     }
 }
